[C-IRIS] move CspaceFreePolytope constructor out of dev folder.
